8.2.4 Добавит логику отображения списка избранного
