+++ project/src/components/favorites/favorites.tsx
@@ -1,13 +1,7 @@
-import { useSelector } from 'react-redux';
-
 import Header from '../header/header';
-import FavoritesCard from '../favorites-card/favorites-card';
-
-import { getOffers } from '../../store/offers/selectors';
+import FavoritesCardsList from '../favorites-cards-list/favorites-cards-list';
 
 function Favorites(): JSX.Element {
-  const offers = useSelector(getOffers);
-
   return (
     <div className="page">
       <Header />
@@ -16,20 +10,12 @@ function Favorites(): JSX.Element {
           <section className="favorites">
             <h1 className="favorites__title">Saved listing</h1>
             <ul className="favorites__list">
-              <li className="favorites__locations-items">
-                <div className="favorites__locations locations locations--current">
-                  <div className="locations__item">
-                    <a className="locations__item-link" href="#">
-                      <span>Amsterdam</span>
-                    </a>
-                  </div>
-                </div>
-                <div className="favorites__places">
-                  {offers
-                    .filter((offer) => offer.isFavorite)
-                    .map((offer) => <FavoritesCard offer={offer} key={offer.id} />)}
-                </div>
-              </li>
+              <FavoritesCardsList city={'Paris'} />
+              <FavoritesCardsList city={'Cologne'} />
+              <FavoritesCardsList city={'Brussels'} />
+              <FavoritesCardsList city={'Amsterdam'} />
+              <FavoritesCardsList city={'Hamburg'} />
+              <FavoritesCardsList city={'Dusseldorf'} />
             </ul>
           </section>
         </div>
